requirements.txt
# Root-level Python dependencies
# Install all with: pip install -r requirements.txt

# Model training (CUDA)
# Install PyTorch with CUDA support:
#   CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#   CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   CPU only:  pip install torch torchvision
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
onnx>=1.15.0

# AWS SDK
boto3>=1.28.0
botocore>=1.31.0

# Deployment scripts
python-dotenv>=1.0.0
requests>=2.31.0

# Backend (ECS)
flask>=3.0.0
gunicorn>=21.2.0
